MetaData post processor fatal
So the code pulls the MetaData information out of the internal content stack prior to doing things like boxed content / list post processing.
But that doesn't reset the keys in the stack. So when the list calculation tries rewinding one level to inject in the list item tags on the prior element, it gets an empty line which isn't ment to happen so it goes boom. (Fixed)